+++ tests/testthat/test-rename-files.R
@@ -1,0 +1,6 @@
+describe("rename_files2()", {
+  it("can rename file if needed.", {
+
+  })
+
+})